cws-status: Exit with code `EX_USAGE` on argument parse error
I tried catching the `argparse.ArgumentError` as described here, but
that didn't seem to work:
https://docs.python.org/3.9/library/argparse.html#exit-on-error

When I wrapped `args = argparser.parse_args()` in a `try` block, it
raised a `SystemExit` exception instead of `argparse.ArgumentError`.

Take inspiration from some of the suggestions in this Stack Overflow
answer:
https://stackoverflow.com/questions/5943249/python-argparse-and-controlling-overriding-the-exit-status-code

Since we just want to change the exit code, call
`argparse.ArgumentParser`'s `error()` method and exit with the error
code we want.